Solve linker issue by placing B64 into a seperate header
